+++ terraform/test-fixtures/apply-provider-configure-disabled/child/main.tf
@@ -1,0 +1,5 @@
+provider "aws" {
+    value = "foo"
+}
+
+resource "aws_instance" "foo" {}
