+++ terraform/test-fixtures/apply-provider-configure-disabled/main.tf
@@ -1,0 +1,7 @@
+provider "aws" {
+    foo = "bar"
+}
+
+module "child" {
+    source = "./child"
+}
